+++ src/plugins/audio.rs
@@ -1,6 +1,9 @@
-use crate::{Level, events::*};
+use crate::{events::*, resources::Level};
 use bevy::prelude::*;
-use bevy_kira_audio::prelude::*;
+use bevy_kira_audio::{
+    AudioApp, AudioControl,
+    prelude::{AudioChannel, AudioPlugin, AudioSource as KiraAudioSource},
+};
 
 #[derive(Resource)]
 pub struct Music;
@@ -14,11 +17,32 @@ pub enum AudioChannelType {
     Sfx,
 }
 
+#[derive(Resource)]
+pub struct AudioThemeHandles {
+    rest: Handle<KiraAudioSource>,
+    combat: Handle<KiraAudioSource>,
+}
+
+impl FromWorld for AudioThemeHandles {
+    fn from_world(world: &mut World) -> Self {
+        let asset_server = world.resource::<AssetServer>();
+        Self {
+            rest: asset_server.load("sfx/loth.ogg"),
+            combat: asset_server.load("sfx/harmony.ogg"),
+        }
+    }
+}
+
 #[allow(dead_code)]
 #[derive(Clone)]
 pub enum AudioCommand {
     Play {
         audio: &'static str,
+        looped: bool,
+        volume: Option<f64>,
+    },
+    PlayHandle {
+        audio: Handle<KiraAudioSource>,
         looped: bool,
         volume: Option<f64>,
     },
@@ -36,6 +60,7 @@ impl Plugin for AudioManagerPlugin {
         app.add_plugins(AudioPlugin)
             .add_audio_channel::<Music>()
             .add_audio_channel::<Sfx>()
+            .init_resource::<AudioThemeHandles>()
             .add_event::<AudioEvent>()
             .add_systems(Update, (play_level_theme, on_audio_event));
     }
@@ -44,11 +69,12 @@ impl Plugin for AudioManagerPlugin {
 fn play_level_theme(
     mut level_changed_events: EventReader<LevelChangedEvent>,
     mut audio_events: EventWriter<AudioEvent>,
+    theme_handles: Res<AudioThemeHandles>,
 ) {
     for event in level_changed_events.read() {
-        let theme_path = match event.new_level {
-            Level::Rest => "sfx/loth.ogg",
-            _ => "sfx/harmony.ogg",
+        let theme_handle = match event.new_level {
+            Level::Rest => theme_handles.rest.clone(),
+            _ => theme_handles.combat.clone(),
         };
 
         // 'flush' the channel before playing the level's theme
@@ -59,8 +85,8 @@ fn play_level_theme(
             },
             AudioEvent {
                 channel: AudioChannelType::Music,
-                command: AudioCommand::Play {
-                    audio: theme_path,
+                command: AudioCommand::PlayHandle {
+                    audio: theme_handle,
                     looped: true,
                     volume: Some(0.15),
                 },
@@ -101,6 +127,21 @@ fn process_audio_command<T>(
             info!("[Audio] Playing: {}", audio);
 
             let mut command = channel.play(asset_server.load(*audio));
+
+            if *looped {
+                command.looped();
+            }
+
+            if let Some(volume) = volume {
+                command.with_volume(*volume);
+            }
+        }
+        AudioCommand::PlayHandle {
+            audio,
+            looped,
+            volume,
+        } => {
+            let mut command = channel.play(audio.clone());
 
             if *looped {
                 command.looped();
